// frontend/src/types/api.ts
export interface ProductDetailResponse {
  id: number;
  name: string;
  slug: string;
  brandName: string;
  brand: {
    id: number;
    name: string;
    slug: string;
  } | null;
  basePrice: string;
  description: string | null;
  specifications: string | null;
  note: string | null;
  colors: {
    id: number;
    name: string;
    hexCode: string;
  }[];
  sizes: string[];
  attributes: {
    name: string;
    value: string[];
  }[];
  variants: {
    id: number;
    sku: string;
    price: string;
    stockQuantity: number;
    colorId: number;
    sizeName: string;
    imageUrls: string[];
  }[];
  categories: {
    id: number;
    name: string;
    slug: string;
  }[];
  audiences: {
    id: number;
    name: string;
    slug: string;
  }[];
  sports: {
    id: number;
    name: string;
    slug: string;
  }[];
}